requirements.txt
Flask
Flask-Restful
Flask-JWT
Flask-SQLAlchemy
Flask-CORS
uwsgi
psycopg2
python-dotenv